requirements_test.txt
-r requirements.txt

## Tests
pytest
pytest-xdist
pytest-cov
pytest-asyncio
pytest-freezegun
coverage

## Utils
ipython
ipdb

## Factories
factory_boy
pytest-factoryboy

## Static Analysis
prospector
